requirements.txt
# K8s AIOps Copilot 依赖

# HolmesGPT 核心
holmesgpt>=0.16.0

# Web 框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# 数据验证
pydantic>=2.0.0

# 配置解析
pyyaml>=6.0

# 控制台输出
rich>=13.0.0

# HTTP 客户端
httpx>=0.25.0

# MCP 支持（可选）
mcp>=0.1.0
starlette>=0.27.0
